requirements.txt
numpy>=1.26
scipy>=1.11
openpyxl>=3.1
PyMuPDF>=1.24
python-docx>=1.1
python-pptx>=0.6.23
Pillow>=10.2
pytest>=8.0
